switch to pthread_everywhere.h, move some platform code to platform.h
