+++ internal/allocator.h
@@ -192,11 +192,7 @@ class Allocator {
  private:
   void DeallocateStorage() {
     assert(!committed_);
-#ifdef _WIN32
-    _aligned_free(storage_);
-#else
-    free(storage_);
-#endif
+    aligned_free(storage_);
     storage_size_ = 0;
   }
