+++ internal/common.h
@@ -18,6 +18,7 @@
 #ifndef GEMMLOWP_INTERNAL_COMMON_H_
 #define GEMMLOWP_INTERNAL_COMMON_H_
 
+#include "../internal/platform.h"
 #include "../profiling/pthread_everywhere.h"
 
 #include <algorithm>
